refactor: 예외 클래스 구체화
