+++ src/main/java/com/example/matchup/matchupbackend/error/exception/InvalidValueEx/InvalidValueException.java
@@ -1,6 +1,7 @@
-package com.example.matchup.matchupbackend.error.exception;
+package com.example.matchup.matchupbackend.error.exception.InvalidValueEx;
 
 import com.example.matchup.matchupbackend.error.ErrorCode;
+import com.example.matchup.matchupbackend.error.exception.CustomException;
 import lombok.Getter;
 
 @Getter
