+++ src/main/java/com/example/matchup/matchupbackend/error/exception/ResourceNotFoundEx/ResourceNotFoundException.java
@@ -1,6 +1,7 @@
-package com.example.matchup.matchupbackend.error.exception;
+package com.example.matchup.matchupbackend.error.exception.ResourceNotFoundEx;
 
 import com.example.matchup.matchupbackend.error.ErrorCode;
+import com.example.matchup.matchupbackend.error.exception.CustomException;
 import lombok.Getter;
 
 @Getter
